backend/pyproject.toml
[project]
name = "marketplace-backend"
version = "0.1.0"
description = "Marketplace Project Workflow System — Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129",
    "sqlmodel>=0.0.33",
    "asyncpg>=0.30",
    "alembic>=1.15",
    "pyjwt[crypto]>=2.11",
    "httpx>=0.28",
    "minio>=7.2",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100
